requirements.txt
Flask
flask-mysqldb
mysqlclient
newsapi-python
python-dotenv